Step 4: Limited Stock
For simplicity sake, "limited stock" in this case is a upper bound on number of apples/oranges that can be in one order.

If I were to change the order service to be constantly running I could have a mock-stock like that such that every order affects it. I could also have implemented a database or cache to track stock. The former wouldn't work with how this service is actually a commandline tool though.

Another way I could have handled this is by having the Mail service handle the stock, and send to that service just what is being ordered. However, this did not make much sense to me since a Mail service should only handle notifications to the user and not item stock.